REFACTOR(Misc): change buttons and cards folders to singular and alter imports, add webkit option to remove white background with autofill [Refs: #34]
