+++ src/components/global/card/CardComp2.vue
