+++ src/components/global/input/GradientInput.vue
@@ -30,9 +30,18 @@ input {
     color: white;
     width: 100%;
     border: 1px solid #535353;
+    
 
     /* border: 1px solid transparent;
     border-image-source: linear-gradient(90deg, #FFFFFF 0%, #999999 100%);
     border-image-slice: 1; */
+}
+
+input:-webkit-autofill,
+input:-webkit-autofill:hover, 
+input:-webkit-autofill:focus, 
+input:-webkit-autofill:active{
+    transition: background-color 5000s ease-in-out 0s;
+    -webkit-text-fill-color: white;
 }
 </style>
